refactor code and delete PollutionJson.java
