+++ src/main/java/dev/entity/Pollution.java
@@ -1,6 +1,8 @@
 package dev.entity;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import dev.enums.QualityAir;
+import dev.dto.PollutionDto;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
@@ -19,47 +21,56 @@ public class Pollution extends BaseEntity {
      * Сoncentration of CO (Carbon monoxide), μg/m3
      */
     @Column(name = "co")
+    @JsonProperty(value = "co")
     private Long carbon;
     /**
      * Сoncentration of NO (Nitrogen monoxide), μg/m3
      */
     @Column(name = "no")
+    @JsonProperty(value = "no")
     private Long nitroMonoxide;
     /**
      * Сoncentration of NO2 (Nitrogen dioxide), μg/m3
      */
     @Column(name = "no2")
+    @JsonProperty(value = "no2")
     private Long nitroDioxide;
     /**
      * Сoncentration of O3 (Ozone), μg/m3
      */
     @Column(name = "o3")
+    @JsonProperty(value = "o3")
     private Long ozone;
     /**
      * Сoncentration of SO2 (Sulphur dioxide), μg/m3
      */
     @Column(name = "so2")
+    @JsonProperty(value = "so2")
     private Long sulphur;
     /**
      * Сoncentration of PM2.5 (Fine particles matter)
      */
     @Column(name = "pm2_5")
+    @JsonProperty(value = "pm2_5")
     private Long fineParticles;
     /**
      * Сoncentration of PM10 (Coarse particulate matter)
      */
     @Column(name = "pm10")
+    @JsonProperty(value = "pm10")
     private Long coarseParticles;
     /**
      * Сoncentration of NH3 (Ammonia), μg/m3
      */
     @Column(name = "nh3")
+    @JsonProperty(value = "nh3")
     private Long ammonia;
     /**
      * Air Quality Index
      */
     @Enumerated
     @Column(name = "air_level")
+    @JsonProperty(value = "air_level")
     private QualityAir airQuality;
 
     /**
@@ -74,7 +85,9 @@ public class Pollution extends BaseEntity {
     public Pollution() {
     }
 
-    public Pollution(Long carbon, Long nitroMonoxide, Long nitroDioxide, Long ozone, Long sulphur, Long fineParticles, Long coarseParticles, Long ammonia, QualityAir airQuality, LocalDateTime date) {
+    public Pollution(Long carbon, Long nitroMonoxide, Long nitroDioxide, Long ozone,
+                     Long sulphur, Long fineParticles, Long coarseParticles, Long ammonia,
+                     QualityAir airQuality, LocalDateTime date) {
         this.carbon = carbon;
         this.nitroMonoxide = nitroMonoxide;
         this.nitroDioxide = nitroDioxide;
@@ -85,86 +98,111 @@ public class Pollution extends BaseEntity {
         this.ammonia = ammonia;
         this.airQuality = airQuality;
         this.date = date;
+    }
+
+    public Pollution(PollutionDto json){
+        PollutionDto.Compos compo = json.getCompo();
+        this
+                .setCarbon(compo.getCarbon())
+                .setNitroMonoxide(compo.getNitroMonoxide())
+                .setNitroDioxide(compo.getNitroDioxide())
+                .setOzone(compo.getOzone())
+                .setSulphur(compo.getSulphur())
+                .setFineParticles(compo.getFineParticles())
+                .setCoarseParticles(compo.getCoarseParticles())
+                .setAmmonia(compo.getAmmonia())
+                .setAirQuality(compo.getAirQuality())
+                .setDate(compo.getDate());
     }
 
     public Long getCarbon() {
         return carbon;
     }
 
-    public void setCarbon(Long carbon) {
+    public Pollution setCarbon(Long carbon) {
         this.carbon = carbon;
+        return this;
     }
 
     public Long getNitroMonoxide() {
         return nitroMonoxide;
     }
 
-    public void setNitroMonoxide(Long nitroMonoxide) {
+    public Pollution setNitroMonoxide(Long nitroMonoxide) {
         this.nitroMonoxide = nitroMonoxide;
+        return this;
     }
 
     public Long getNitroDioxide() {
         return nitroDioxide;
     }
 
-    public void setNitroDioxide(Long nitroDioxide) {
+    public Pollution setNitroDioxide(Long nitroDioxide) {
         this.nitroDioxide = nitroDioxide;
+        return this;
     }
 
     public Long getOzone() {
         return ozone;
     }
 
-    public void setOzone(Long ozone) {
+    public Pollution setOzone(Long ozone) {
         this.ozone = ozone;
+        return this;
     }
 
     public Long getSulphur() {
         return sulphur;
     }
 
-    public void setSulphur(Long sulphur) {
+    public Pollution setSulphur(Long sulphur) {
         this.sulphur = sulphur;
+        return this;
     }
 
     public Long getFineParticles() {
         return fineParticles;
     }
 
-    public void setFineParticles(Long fineParticles) {
+    public Pollution setFineParticles(Long fineParticles) {
         this.fineParticles = fineParticles;
+        return this;
     }
 
     public Long getCoarseParticles() {
         return coarseParticles;
     }
 
-    public void setCoarseParticles(Long coarseParticles) {
+    public Pollution setCoarseParticles(Long coarseParticles) {
         this.coarseParticles = coarseParticles;
+        return this;
     }
 
     public Long getAmmonia() {
         return ammonia;
     }
 
-    public void setAmmonia(Long ammonia) {
+    public Pollution setAmmonia(Long ammonia) {
         this.ammonia = ammonia;
+        return this;
     }
 
     public QualityAir getAirQuality() {
         return airQuality;
     }
 
-    public void setAirQuality(QualityAir airQuality) {
+    public Pollution setAirQuality(QualityAir airQuality) {
         this.airQuality = airQuality;
+        return this;
     }
 
     public LocalDateTime getDate() {
         return date;
     }
 
-    public void setDate(LocalDateTime date) {
+    public Pollution setDate(LocalDateTime date) {
         this.date = date;
+        return this;
     }
 }
